refactor: move logic from ProductController to ProductService

+++ src/main/java/com/estudo/primeiro/crud/services/ProductService.java
@@ -1,10 +1,12 @@
 package com.estudo.primeiro.crud.services;
 
+import com.estudo.primeiro.crud.dtos.ProductDTO;
 import com.estudo.primeiro.crud.entities.Product;
 import com.estudo.primeiro.crud.repositories.ProductRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -14,21 +16,45 @@ public class ProductService {
         this.repository = repository;
     }
 
-    public List<Product> findAll(){
-        return repository.findAll();
+    public ProductDTO Create(ProductDTO dto) {
+        Product product = dto.MapToEntity();
+        repository.save(product);
+
+        return product.MapToDTO();
     }
 
-    public Product findById(long id){
-        return repository.findById(id).orElse(null);
+    public List<ProductDTO> FindAll(){
+        return repository.findAll().stream().map(Product::MapToDTO).toList();
     }
 
-    public Product save(Product product){
-        return repository.save(product);
-    }
-
-    public void delete(Long id){
-       repository.deleteById(id);
+    public Optional<ProductDTO> FindById(long id){
+        return repository.findById(id).map(Product::MapToDTO);
     }
 
 
+    public void Delete(Long id){
+        Optional<Product> optionalProduct = repository.findById(id);
+
+        if (optionalProduct.isEmpty()) {
+            // TODO: criar mensagem de erro
+            return;
+        }
+
+        repository.deleteById(id);
+    }
+
+    public ProductDTO Update(ProductDTO dto, Long id) {
+        Optional<Product> optionalProduct = repository.findById(id);
+
+        if (optionalProduct.isEmpty()) {
+            // TODO: criar mensagem de erro
+            return null;
+        }
+
+        Product product = optionalProduct.get();
+        product.UpdateProduct(dto.getName(), dto.getDescription(), dto.getPrice());
+        repository.save(product);
+
+        return product.MapToDTO();
+    }
 }
